Resolved some of the changes requested in PR #34
