+++ src/App.tsx
@@ -9,8 +9,7 @@ import {Navigation} from './components/elements/nav/Main';
 import {AnonymousRoute} from './components/elements/routes/AnonymousRoute';
 import {PrivateRoute} from './components/elements/routes/PrivateRoute';
 import {PublicRoute} from './components/elements/routes/PublicRoute';
-// import PurchaseModal from './components/elements/table/PurchaseModal';
-import {RestockPurchasesTable} from './components/elements/table/RestockTableMain';
+import RestockPurchasesTable from './components/elements/table/RestockPurchasesTable';
 import {Authenticated} from './components/pages/Authenticated';
 import {ForgotPassword} from './components/pages/ForgotPassword';
 import {Homepage} from './components/pages/Homepage';
