+++ src/const/values.ts
@@ -1,0 +1,7 @@
+/**
+ * Const values of the app.
+ */
+export default class AppValues {
+  // Restock Purchases
+  static TAX_RATE = 0.07;
+}
