+++ src/state/restock/restockData.ts
@@ -1,21 +1,36 @@
+export type RestockType = 'Manual' | 'Auto';
+
+export type RestockStatusType = 'processing' | 'shipped' | 'completed';
+
+/**
+ * The base fields for every restock purchase.
+ */
 export type RestockItemInfo = {
     id: string;
-    status: string;
+    status: RestockStatusType;
     purchaseDate: string;
     totalPrice: number;
-    type: 'Manual' | 'Auto';
+    type: RestockType;
 }
 
+/**
+ * The base fields for every item purchased during
+ * a restock purchase.
+ */
 export type ItemDataInfo = {
     description: string;
     unitPrice: number;
     quantity: number;
 }
 
+/**
+ * A collection of all the items purchased during a
+ * restock purchase
+ */
 export type Items = {
     itemsData: Array<ItemDataInfo>
 }
 
-export type RestockDataInfo = RestockItemInfo & Items;
+export type RestockDataEntry = RestockItemInfo & Items;
 
-export type RestockData = Array<RestockDataInfo>;
+export type RestockData = Array<RestockDataEntry>;
