+++ src/state/restock/restockName.ts
@@ -2,8 +2,7 @@ export const RESTOCK_STATE_NAME = 'restock';
 
 export enum RestockDispatcherName {
     // TODO: SHOULD BE CONNECTED TO THE ACCOUNT PAGE
-    EDIT_RESTOCK_SETTINGS = 'editRestockSettings',
+    UPDATE_RESTOCK_SETTINGS = 'updateRestockSettings',
     FETCH_RESTOCK_DATA = 'fetchRestockData',
-    // TODO: NOT SURE IF THIS SHOULD BE AN ACTION
     SORT_RESTOCK_DATA = 'sortRestockData',
 }
