+++ src/state/restock/restockState.ts
@@ -2,5 +2,5 @@ import {StateBase} from '../base';
 import {RestockData} from './restockData';
 
 export type RestockState = StateBase & {
-    restockData: RestockData | null;
+    restockData: RestockData | [];
 }
